requirements.txt
Package     Version
----------- -------
lxml        5.1.0
pillow      10.2.0
pip         22.3.1
python-pptx 0.6.23
setuptools  65.5.1
structlog   24.1.0
wheel       0.38.4
XlsxWriter  3.2.0

structlog~=24.1.0